Replace Indicator with Radar chart
