+++ src/components/charts/IypGenericRadarChart.vue
@@ -1,0 +1,78 @@
+<script setup>
+import ReactiveChart from './ReactiveChart.vue'
+import { ref, onMounted, watch } from 'vue'
+
+const props = defineProps({
+  chartData: {
+    type: Array,
+    default: () => [],
+  },
+  chartLayout: {
+    type: Object,
+    default: () => ({}),
+  },
+  config: {
+    type: Object,
+    default: () => ({}),
+  }
+})
+
+const localChartData = ref([])
+const actualChartData = ref([])
+const actualChartLayout = ref({})
+
+const renderChart = () => {
+  let data = formatChartData(localChartData.value)
+
+  let layout = {
+    ...props.chartLayout,
+  }
+
+  actualChartData.value = data
+  actualChartLayout.value = layout
+}
+
+const formatChartData = (arrayOfObjects) => {
+  if (!arrayOfObjects || arrayOfObjects.length === 0) {
+    return []
+  }
+
+  let data = []
+  if (arrayOfObjects.length) {
+    data = [
+      {
+        type: 'scatterpolar',
+        r: arrayOfObjects.map(val => val.r),
+        theta: arrayOfObjects.map(val => val.theta),
+        fill: 'toself',
+        ...props.config
+      }
+    ]
+  }
+
+  return data
+}
+
+const init = () => {
+  if (props.chartData && props.chartData.length > 0) {
+    localChartData.value = props.chartData
+    renderChart()
+  }
+}
+
+watch(() => props.chartData, () => {
+  init()
+}, { deep: true })
+
+onMounted(() => {
+  init()
+})
+</script>
+
+<template>
+  <ReactiveChart
+    :layout="actualChartLayout"
+    :traces="actualChartData"
+    :chart-title="actualChartLayout && actualChartLayout.title"
+  />
+</template>
